Split ScaledTimerMinimum into its own file
